Move all socket lookup and defaulting logic to the same place
